Move Slider into PlayerControls directory
