+++ packages/studio-base/src/components/PlaybackControls/Slider.tsx
